Move HTTPClient and FeedItemsMapper to its own file
