+++ EssentialFeed/EssentialFeed/Feed API/RemoteFeedLoader.swift
@@ -7,11 +7,6 @@
 
 import Foundation
 
-
-public protocol HTTPClient {
-    func get(from url: URL, completion: @escaping(Result<(Data, HTTPURLResponse), Error>) -> Void)
-}
-
 public final class RemoteFeedLoader {
     private let url: URL
     private let client: HTTPClient
@@ -32,7 +27,7 @@ public final class RemoteFeedLoader {
             case let .success((data, response)):
 
                 do {
-                    let items = try FeedItemMapper.map(data, response)
+                    let items = try FeedItemsMapper.map(data, response)
                     completion(.success(items))
                 } catch {
                     completion(.failure(.invalidData))
@@ -44,33 +39,3 @@ public final class RemoteFeedLoader {
     }
 }
 
-private class FeedItemMapper {
-    private struct Root: Decodable {
-        let items: [Item]
-    }
-
-    private struct Item: Decodable {
-        let id: UUID
-        let description: String?
-        let location: String?
-        let image: URL
-        
-        var item: FeedItem {
-            return FeedItem(id: id,
-                            description: description,
-                            location: location,
-                            imageURL: image)
-        }
-    }
-    
-    static var OK_200: Int {return 200}
-    
-    static func map(_ data: Data, _ response: HTTPURLResponse) throws -> [FeedItem] {
-        guard response.statusCode == OK_200 else {
-            throw RemoteFeedLoader.Error.invalidData
-        }
-        
-        let root = try JSONDecoder().decode(Root.self, from: data)
-        return root.items.map{ $0.item }
-    }
-}
